Added surfacing of fluid"
The algorithm is as follows: first, grid_mark_cells now estimates the fraction of a cell that is occupied by fluid. This is done using the same bilinear kernel that is used for transferring velocities to the grid.
It is then observed that the fluid fraction value acts as a crude estimate of the location of the free surface within the cell - a fraction of 0.5 indicates it's in the centre, 0 or 1 and it passes very close to the edge. So to generate the surface, marching squares is applied using 0.5 as the isovalue.